+++ modules/database/rds/stack.py
@@ -43,10 +43,12 @@ class RDSDatabaseStack(cdk.Stack):
         subnet_ids: list[str],
         engine: str,
         engine_version: str,
+        database_name: str,
         username: str,
         credential_rotation_days: int,
         instance_type: str,
         removal_policy: cdk.RemovalPolicy,
+        is_accessible_from_vpc: bool = False,
         port: int | None = None,
         stack_description: str | None = None,
         **kwargs: Any,
@@ -104,6 +106,7 @@ class RDSDatabaseStack(cdk.Stack):
             scope=self,
             id="RDS Database",
             port=port,
+            database_name=database_name,
             credentials=rds.Credentials.from_secret(self.db_credentials_secret),
             engine=_get_db_instance_engine(engine, engine_version),
             instance_type=ec2.InstanceType(instance_type),
@@ -115,11 +118,12 @@ class RDSDatabaseStack(cdk.Stack):
             removal_policy=removal_policy,
         )
 
-        # Adds an ingress rule which allows resources in the VPC's CIDR to access the database.
-        self.database.connections.allow_default_port_from(
-            other=ec2.Peer.ipv4("10.0.0.0/24"),
-            description="Allows resources in the VPC CIDR to access the database",
-        )
+        if is_accessible_from_vpc:
+            # Adds an ingress rule which allows resources in the VPC's CIDR to access the database.
+            self.database.connections.allow_default_port_from(
+                other=ec2.Peer.ipv4(vpc.vpc_cidr_block),
+                description="Allows resources in the VPC CIDR to access the database",
+            )
 
         # Set up CDK nag
         cdk.Aspects.of(self).add(cdk_nag.AwsSolutionsChecks(log_ignores=True))